// src/types/clipper.d.ts
export interface WhisperWord {
    word: string;
    start: number;
    end: number;
    confidence: number;
}